# openapi/openapi_ire.yaml
openapi: 3.0.3
info:
  title: Identity Resolution Engine API
  version: 1.1.0
  description: >
    Identity Resolution Engine MVP API with demo (CSV/in-memory) mode and Oracle persistence mode.
paths:
  /health:
    get:
      summary: API health check
      operationId: health
      responses:
        "200":
          description: API is available
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  service: { type: string, example: ire-api }
                  mode: { type: string, enum: [demo, oracle] }

  /ingest:
    post:
      summary: Ingest source records and optionally run matching
      operationId: ingest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequest"
      responses:
        "200":
          description: Source record accepted and persisted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  source_name: { type: string }
                  source_pk: { type: string }
                  source_record_id: { type: integer }
                  persistence_mode: { type: string, enum: [demo, oracle] }
                  normalized:
                    $ref: "#/components/schemas/NormalizedFields"
                  match:
                    $ref: "#/components/schemas/MatchEnvelope"

  /match:
    post:
      summary: Match an incoming record against golden records
      operationId: match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequest"
      responses:
        "200":
          description: Matching decision and evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"

  /merge:
    post:
      summary: Merge incoming data into a golden record and write merge/link audit entries
      operationId: merge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeRequest"
      responses:
        "200":
          description: Merged golden record
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  golden_id: { type: string }
                  persistence_mode: { type: string, enum: [demo, oracle] }
                  golden_record:
                    $ref: "#/components/schemas/PersonRecord"
        "404":
          description: Golden record not found

  /golden/{golden_id}:
    get:
      summary: Retrieve a golden record by ID
      operationId: getGolden
      parameters:
        - name: golden_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Golden record
          content:
            application/json:
              schema:
                type: object
                properties:
                  golden_id: { type: string }
                  record:
                    $ref: "#/components/schemas/PersonRecord"
        "404":
          description: Golden record not found

  /review/tasks:
    get:
      summary: List open manual review tasks
      operationId: listReviewTasks
      responses:
        "200":
          description: Open manual review tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReviewTask"

  /review/{task_id}/decision:
    post:
      summary: Record reviewer decision for a manual review task
      operationId: submitReviewDecision
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  task_id: { type: integer }
                  decision: { type: string }
                  task_status: { type: string }
                  link_created: { type: boolean }
        "404":
          description: Task not found

components:
  schemas:
    PersonRecord:
      type: object
      properties:
        record_id: { type: string, nullable: true }
        source_type: { type: string, nullable: true }
        name: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        hkid: { type: string, nullable: true }
        emplid: { type: string, nullable: true }
        studentid: { type: string, nullable: true }
        alumniid: { type: string, nullable: true }
        address: { type: string, nullable: true }

    NormalizedFields:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }

    Candidate:
      type: object
      properties:
        golden_id: { type: string, nullable: true }
        confidence: { type: number, format: float }
        reason: { type: string }
        sims:
          type: object
          additionalProperties:
            type: number

    MatchEnvelope:
      type: object
      properties:
        decision: { type: string }
        reason: { type: string }
        confidence: { type: number, format: float }
        best_match_record_id: { type: string, nullable: true }
        sims:
          type: object
          additionalProperties:
            type: number
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
        manual_review_task_id: { type: integer }

    IngestRequest:
      type: object
      required: [source_name, source_pk, payload]
      properties:
        source_name: { type: string }
        source_pk: { type: string }
        payload:
          $ref: "#/components/schemas/PersonRecord"
        run_match: { type: boolean, default: false }
        ingest_user: { type: string, default: ingest_service }

    MatchRequest:
      type: object
      required: [incoming]
      properties:
        incoming:
          $ref: "#/components/schemas/PersonRecord"
        golden_records:
          type: array
          items:
            $ref: "#/components/schemas/PersonRecord"
          description: Optional explicit golden candidate set.

    MatchResponse:
      type: object
      properties:
        decision:
          type: string
          enum: [auto_merge, manual_review, create_new_golden_record]
        confidence:
          type: number
          format: float
        best_match_index:
          type: integer
          nullable: true
        best_match_record_id:
          type: string
          nullable: true
        similarities:
          type: object
          additionalProperties:
            type: number
        explanation:
          type: string
        reason:
          type: string
        evidence:
          type: object
          additionalProperties:
            type: number
        sims:
          type: object
          additionalProperties:
            type: number
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"

    MergeRequest:
      type: object
      required: [golden_id, incoming]
      properties:
        golden_id: { type: string }
        incoming:
          $ref: "#/components/schemas/PersonRecord"
        source_record_id: { type: integer, nullable: true }
        confidence: { type: number, format: float, default: 1.0 }
        link_method: { type: string, default: manual }
        actor: { type: string, default: system }
        evidence:
          type: object
          additionalProperties:
            type: number

    ReviewTask:
      type: object
      properties:
        task_id: { type: integer }
        source_record_id: { type: integer }
        candidate_goldens:
          type: array
          items: { type: string }
        best_confidence: { type: number, format: float, nullable: true }
        status: { type: string }
        assigned_to: { type: string, nullable: true }
        reviewer_decision: { type: string, nullable: true }
        notes: { type: string, nullable: true }

    ReviewDecisionRequest:
      type: object
      required: [decision, reviewer_id]
      properties:
        decision:
          type: string
          enum: [accept, merge, reject, new, escalate]
        reviewer_id: { type: string }
        notes: { type: string, nullable: true }
        golden_id: { type: string, nullable: true }